remove availability arbitration scenario
