+++ monitoring/uss_qualifier/scenarios/astm/utm/dss/__init__.py
@@ -1,4 +1,4 @@
 from .subscription_validation import SubscriptionValidation
 from .subscription_simple import SubscriptionSimple
 from .crdb_access import CRDBAccess
-from .availability_arbitration import AvailabilityArbitration
+from .report import Report
