+++ monitoring/uss_qualifier/suites/astm/utm/dss_probing.yaml
@@ -53,6 +53,6 @@ actions:
       resources:
         crdb_cluster: dss_crdb_cluster
   - test_scenario:
-      scenario_type: scenarios.astm.utm.dss.AvailabilityArbitration
+      scenario_type: scenarios.astm.utm.dss.Report
       resources:
         dss: dss
